Student side navbar component and logout
Created a separate component for the student side navbar with all functionality intact. Implemented the basic logout functionality. Removed all persisting lint errors
